update
- pytest
- re-arrange the function in the utils
