+++ src/utils/common_utils.py
@@ -229,3 +229,6 @@ def load_dict_from_json(filepath: Path) -> dict:
 def generate_explode(nb_categories: int) -> list:
     explode = [1 / nb_categories for k in range(nb_categories)]
     return explode
+
+
+
